Move MONTHS inside of the HHData step folder
